abuse2sqlite/requirements.txt
requests
argparse
progressbar2
maxminddb
sqlite3
queue
threading
